feat: refactor community page to use base wrapper
